0.2

+++ SnakeBite/formMain.cs
@@ -4,9 +4,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Windows.Forms;
-using System.Xml.Serialization;
-using SnakeBite.GzsTool;
-using GzsTool.Utility;
 
 namespace SnakeBite
 {
@@ -22,10 +19,11 @@ namespace SnakeBite
 
         private void formMain_Load(object sender, EventArgs e)
         {
+            labelVersion.Text = Application.ProductVersion;
             // check if user has specified valid install path
             if (!ModManager.ValidInstallPath)
             {
-                MessageBox.Show("Please locate your MGSV installation to continue.", "SnakeBite", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                MessageBox.Show("Please locate your MGSV installation to continue. If this is your first time running SnakeBite, it is recommended that you reset your MGSV installation before continuing.", "SnakeBite", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
                 this.Show();
                 tabControl.SelectedIndex = 1;
                 buttonFindMGSV_Click(null, null);
@@ -40,18 +38,13 @@ namespace SnakeBite
 
             this.Show();
 
-            // hash 01.dat and check against last known hash
-            // if hash doesn't match, prompt user to regen game file database
-            string datHash = ModManager.HashFile(ModManager.GameArchivePath);
-            if(datHash != objSettings.GameData.DatHash)
+            // check hash for dat file, if changed, rebuild database
+            if (!ModManager.CheckDatHash())
             {
-                MessageBox.Show("Game data modified outside of SnakeBite. Game cache must be rebuilt and previously installed mods will no longer appear.",
+                MessageBox.Show("Game data modified outside of SnakeBite. SnakeBite will now attempt to recache game data",
                                 "Game data hash mismatch", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 buttonBuildGameDB_Click(null, null);
-                objSettings.GameData.DatHash = ModManager.HashFile(ModManager.GameArchivePath);
-                objSettings.SaveSettings();
             }
-
         }
 
         private void LoadInstalledMods(bool resetSelection = false)
@@ -67,7 +60,7 @@ namespace SnakeBite
 
             listInstalledMods.Items.Clear();
 
-            if(objSettings.ModEntries.Count > 0)
+            if (objSettings.ModEntries.Count > 0)
             {
                 panelModDetails.Visible = true;
                 labelNoMods.Visible = false;
@@ -86,34 +79,32 @@ namespace SnakeBite
                     {
                         listInstalledMods.SelectedIndex = -1;
                     }
-
                 }
-            } else
+            }
+            else
             {
                 panelModDetails.Visible = false;
                 labelNoMods.Visible = true;
             }
-
-            
         }
 
         private void buttonUninstallMod_Click(object sender, EventArgs e)
         {
+            if(!(listInstalledMods.SelectedIndex >= 0)) return;
+            
             // Get selected mod
             ModEntry mod = objSettings.ModEntries[listInstalledMods.SelectedIndex];
+            if (!(MessageBox.Show(String.Format("{0} will be uninstalled.", mod.Name), "SnakeBite", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == DialogResult.OK)) return;
 
-            showProgressWindow(String.Format("Uninstalling {0}...", mod.Name));
 
-            // Uninstall mod
-            ModManager.UninstallMod(mod);
+            showProgressWindow(String.Format("Please wait while {0} is uninstalled...", mod.Name));
 
             // Remove from mod database
             objSettings.ModEntries.Remove(mod);
             objSettings.SaveSettings();
 
-            // Update DAT hash
-            Properties.Settings.Default.DatHash = ModManager.HashFile(ModManager.GameArchivePath);
-            Properties.Settings.Default.Save();
+            // Uninstall mod
+            ModManager.UninstallMod(mod);
 
             // Update installed mod list
             LoadInstalledMods(true);
@@ -136,46 +127,66 @@ namespace SnakeBite
             modMetadata.ReadFromFile("metadata.xml");
             File.Delete("metadata.xml"); // delete temp metadata
 
-            // search installed mods for conflicts
-            List<string> conflictingMods = new List<string>();
-
-            foreach (ModEntry modEntry in objSettings.ModEntries) // iterate through installed mods
+            if(!checkConflicts.Checked)
             {
-                foreach (ModQarEntry qarEntry in modMetadata.ModQarEntries) // iterate qar files from new mod
-                {
-                    ModQarEntry conflicts = modEntry.ModQarEntries.FirstOrDefault(entry => entry.FilePath == qarEntry.FilePath);
-                    if (conflicts != null)
-                    {
-                        conflictingMods.Add(modEntry.Name);
-                        break;
-                    }
-                }
+                // search installed mods for conflicts
+                List<string> conflictingMods = new List<string>();
 
-                foreach (ModFpkEntry fpkEntry in modMetadata.ModFpkEntries) // iterate fpk files from new mod
+                foreach (ModEntry modEntry in objSettings.ModEntries) // iterate through installed mods
                 {
-                    ModFpkEntry conflicts = modEntry.ModFpkEntries.FirstOrDefault(entry => entry.FpkFile == fpkEntry.FpkFile && entry.FilePath == fpkEntry.FilePath);
-                    if (conflicts != null)
+                    foreach (ModQarEntry qarEntry in modMetadata.ModQarEntries) // iterate qar files from new mod
                     {
-                        if (!conflictingMods.Contains(modEntry.Name))
+                        ModQarEntry conflicts = modEntry.ModQarEntries.FirstOrDefault(entry => entry.FilePath == qarEntry.FilePath);
+                        if (conflicts != null)
                         {
                             conflictingMods.Add(modEntry.Name);
                             break;
                         }
                     }
-                }
-            }
 
-            // if the mod conflicts, display message
-            if (conflictingMods.Count > 0)
-            {
-                string msgboxtext = "The selected mod conflicts with these mods:\n";
-                foreach (string Conflict in conflictingMods)
-                {
-                    msgboxtext += Conflict + "\n";
+                    foreach (ModFpkEntry fpkEntry in modMetadata.ModFpkEntries) // iterate fpk files from new mod
+                    {
+                        ModFpkEntry conflicts = modEntry.ModFpkEntries.FirstOrDefault(entry => entry.FpkFile == fpkEntry.FpkFile && entry.FilePath == fpkEntry.FilePath);
+                        if (conflicts != null)
+                        {
+                            if (!conflictingMods.Contains(modEntry.Name))
+                            {
+                                conflictingMods.Add(modEntry.Name);
+                                break;
+                            }
+                        }
+                    }
                 }
-                msgboxtext += "\nPlease uninstall the mods above and try again.";
-                MessageBox.Show(msgboxtext, "Installation Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                return;
+
+                // if the mod conflicts, display message
+                if (conflictingMods.Count > 0)
+                {
+                    string msgboxtext = "The selected mod conflicts with these mods:\n";
+                    foreach (string Conflict in conflictingMods)
+                    {
+                        msgboxtext += Conflict + "\n";
+                    }
+                    msgboxtext += "\nPlease uninstall the mods above and try again.";
+                    MessageBox.Show(msgboxtext, "Installation Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+
+                bool sysConflict = false;
+                // check for system file conflicts
+                foreach (ModQarEntry gameQarFile in objSettings.GameData.GameQarEntries)
+                {
+                    if (modMetadata.ModQarEntries.Count(entry => entry.FilePath == gameQarFile.FilePath) > 0) sysConflict = true;
+                }
+
+                foreach (ModFpkEntry gameFpkFile in objSettings.GameData.GameFpkEntries)
+                {
+                    if (modMetadata.ModFpkEntries.Count(entry => entry.FilePath == gameFpkFile.FilePath && entry.FpkFile == gameFpkFile.FpkFile) > 0) sysConflict = true;
+                }
+                if (sysConflict)
+                {
+                    MessageBox.Show("The selected mod conflicts with existing MGSV system files.", "SnakeBite", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
             }
 
             DialogResult confirmInstall = MessageBox.Show(String.Format("You are about to install {0}, continue?", modMetadata.Name), "SnakeBite", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
@@ -183,16 +194,14 @@ namespace SnakeBite
 
             showProgressWindow(String.Format("Installing {0}, please wait...", modMetadata.Name));
 
-            // Install mod to 01.dat
-            ModManager.InstallMod(openModFile.FileName);
-
-            // Update DAT hash
-            objSettings.GameData.DatHash = ModManager.HashFile(ModManager.GameArchivePath);
-            objSettings.SaveSettings();
 
             // Install mod to game database
             objSettings.ModEntries.Add(modMetadata);
             objSettings.SaveSettings();
+
+            // Install mod to 01.dat
+            ModManager.InstallMod(openModFile.FileName);
+
             LoadInstalledMods();
             listInstalledMods.SelectedIndex = listInstalledMods.Items.Count - 1;
 
@@ -213,7 +222,6 @@ namespace SnakeBite
                 textDescription.Text = selectedMod.Description;
             }
         }
-
 
         private void buttonFindMGSV_Click(object sender, EventArgs e)
         {
@@ -239,39 +247,62 @@ namespace SnakeBite
 
         private void hideProgressWindow()
         {
-
             this.Enabled = true;
             progWindow.Hide();
-            
         }
 
         private void buttonBuildGameDB_Click(object sender, EventArgs e)
         {
-            // clears all mods, resets game data files
-            if(sender!=null)
+            // attempt to determine which files are game files and which are mod files
+            if (sender != null)
             {
-                DialogResult areYouSure = MessageBox.Show("Currently installed mods will be merged with base game data and you will no longer be able to uninstall them. Are you sure?", "SnakeBite", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                DialogResult areYouSure = MessageBox.Show("SnakeBite will recache the installed game data and remove any invalid mods. This may take some time.\n\nContinue?", "SnakeBite", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
                 if (areYouSure == DialogResult.No) return;
             }
 
             showProgressWindow("Rebuilding game data cache...");
 
-            objSettings.GameData = ModManager.BuildGameData();
-            objSettings.ModEntries = new List<ModEntry>();
+            objSettings.GameData = ModManager.RebuildGameData();
             objSettings.SaveSettings();
 
-            LoadInstalledMods();
+            ModManager.CleanupModSettings();
+
+            ModManager.UpdateDatHash();
+
+            LoadInstalledMods(true);
 
             hideProgressWindow();
-            tabControl.SelectedIndex = 0;
+
+            if(sender!=null) tabControl.SelectedIndex = 0;
         }
 
         private void checkConflicts_CheckedChanged(object sender, EventArgs e)
         {
-            if(checkConflicts.Checked)
+            if (checkConflicts.Checked)
             {
                 MessageBox.Show("Disabling conflict checking will allow SnakeBite to overwrite existing files, potentially causing unpredictable results. It is recommended you backup 01.dat before continuing!", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
+        }
+
+        private void labelModAuthor_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void buttonLaunch_Click(object sender, EventArgs e)
+        {
+            System.Diagnostics.Process.Start("steam://run/287700/");
+            Application.Exit();
+        }
+
+        private void linkLabel1_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
+        {
+            System.Diagnostics.Process.Start(this.linkLabel1.Text);
+        }
+
+        private void label2_Click(object sender, EventArgs e)
+        {
+
         }
     }
 }
